Changed the TrainModel file location.
