--- src/application/TrainModel/TrainControllerInterface.java
@@ -1,7 +1,0 @@
-package application.TrainModel;
-
-public interface TrainControllerInterface {
-
-    float getPower();
-    float getBrake();
-}
