--- src/application/TrainModel/TrainModelInterface.java
@@ -1,8 +1,0 @@
-package application.TrainModel;
-
-public interface TrainModelInterface {
-
-    TrainModel getTrain(String trainID);
-
-    boolean dispatchTrain(String trainID);
-}
