--- src/application/TrainModel/train_model_FX.fxml
@@ -1,42 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.CheckBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.control.ListView?>
-<?import javafx.scene.control.ToggleButton?>
-<?import javafx.scene.image.ImageView?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.ColumnConstraints?>
-<?import javafx.scene.layout.GridPane?>
-<?import javafx.scene.layout.RowConstraints?>
-<?import javafx.scene.layout.VBox?>
-
-<?import javafx.scene.control.TableView?>
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.TrainModel.TrainModelFX">
-   <children>
-      <GridPane prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-        <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-          <ColumnConstraints hgrow="SOMETIMES" minWidth="10.0" prefWidth="100.0" />
-        </columnConstraints>
-        <rowConstraints>
-          <RowConstraints maxHeight="196.0" minHeight="10.0" prefHeight="90.0" vgrow="SOMETIMES" />
-          <RowConstraints maxHeight="335.0" minHeight="10.0" prefHeight="310.0" vgrow="SOMETIMES" />
-        </rowConstraints>
-         <children>
-            <TableView fx:id="trainInfo" prefHeight="276.0" prefWidth="313.0" GridPane.rowIndex="1" />
-            <ImageView fitHeight="86.0" fitWidth="313.0" pickOnBounds="true" preserveRatio="true" />
-            <ToggleButton mnemonicParsing="false" text="Emergency Brake" GridPane.columnIndex="1" />
-            <VBox prefHeight="276.0" prefWidth="200.0" GridPane.columnIndex="1" GridPane.rowIndex="1">
-               <children>
-                  <Label text="Faults" />
-                  <CheckBox mnemonicParsing="false" text="Engine Failure" />
-                  <CheckBox mnemonicParsing="false" text="Brake Failure" />
-                  <CheckBox mnemonicParsing="false" text="Door Failure" />
-                  <CheckBox mnemonicParsing="false" text="Signal Pickup Failure" />
-               </children>
-            </VBox>
-         </children>
-      </GridPane>
-   </children>
-</AnchorPane>
