--- src/application/TrainModel/train_model_main_FX.fxml
@@ -1,24 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-
-<?import javafx.scene.control.Button?>
-<?import javafx.scene.control.ChoiceBox?>
-<?import javafx.scene.control.Label?>
-<?import javafx.scene.layout.AnchorPane?>
-<?import javafx.scene.layout.BorderPane?>
-<?import javafx.scene.layout.VBox?>
-
-<AnchorPane prefHeight="400.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/8.0.172-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="application.TrainModel.TrainModelMainFX">
-   <children>
-      <BorderPane prefHeight="400.0" prefWidth="600.0" AnchorPane.bottomAnchor="0.0" AnchorPane.leftAnchor="0.0" AnchorPane.rightAnchor="0.0" AnchorPane.topAnchor="0.0">
-         <center>
-            <VBox BorderPane.alignment="CENTER">
-               <children>
-                  <Label text="Train Model Main" />
-                  <ChoiceBox id="trainSelector" fx:id="trainSelectorModel" onMouseClicked="#selectTrain" prefWidth="150.0" />
-                  <Button id="trainSelectButton" mnemonicParsing="false" onMouseClicked="#clickSelect" text="Select" />
-               </children>
-            </VBox>
-         </center>
-      </BorderPane>
-   </children>
-</AnchorPane>
